fix: adjust Server page to use server side auth fetching, adjust Server page text to include RSC, move components folder to root dir
